requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
audioread==2.1.9
backcall==0.1.0
bleach==3.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.0
decorator==4.4.1
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
gast==0.3.3
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.9.0
idna==2.9
imageio==2.6.1
imageio-ffmpeg==0.4.3
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
munch==2.5.0
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numba==0.48.0
numpy==1.18.4
olefile==0.46
opencv-python==3.4.2.17
opt-einsum==3.1.0
paddlepaddle-gpu==2.0.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
ppgan==2.0.0b0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.1
ptyprocess==0.6.0
PyAudio==0.2.11
pycparser==2.20
Pygments==2.5.2
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
resampy==0.2.2
scikit-image==0.17.2
scikit-learn==0.24.1
scikit-multilearn==0.2.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.13.0
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
tensorboard==2.0.0
tensorboardX==2.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.12.8
toolz==0.10.0
torch==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.4.0
tornado==6.0.3
tqdm==4.46.0
traitlets==4.3.3
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
